pyproject.toml
[project]
name = "mcp_databricks_filtering"
version = "0.0.1"
description = "Tag-based table filtering for the Databricks MCP server"
requires-python = ">=3.10,<3.13"
dependencies = [
    "databricks-sdk",
    "sqlglot",
]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
    "databricks-connect>=15.4,<15.5",
]

[project.scripts]
mcp-table-filter = "mcp_databricks_filtering.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
markers = [
    "online: tests that require a live Databricks connection",
]
